Split the code in `length.rs` into two modules
